Required files added
Added mysqlconnection.py to connect the flask application and the MySQL database.
Added server.py with 2 routes. '/': To login/register and '/success': to confirm the login process has been completed with no errors
Added index.html with two forms and a link. One form to login and the other to register. The link is to be used only for testing. May use for a different reason later.
Added success.html with a header confirming success of login and a link for testing. This link also may be used in the future for something else.