Styling/Salt/logout/DB
Added styling to both templates so it didn't look too generic(styling was just random not with effort).
Added salt to registration so logging would draw on that to decrypt the password in the db.
Added log out forms on both templates and links for css.
Added DB used/created for this assignment.